+++ R/ifcb_which_basin.R
@@ -89,8 +89,8 @@ ifcb_which_basin <- function(latitudes, longitudes, plot = FALSE, shape_file = N
 
   # Return a logical vector indicating whether each point is in a basin, or the plot if requested
   if (plot) {
-    return(plot_obj)
+    plot_obj
   } else {
-    return(basin_names)
+    basin_names
   }
 }