+++ NEWS.md
@@ -1,3 +1,5 @@
+# iRfcb (development version)
+
 # iRfcb 0.5.0
 
 ## New features